requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
Deprecated==1.2.18
google-auth==2.38.0
google-genai==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGithub==2.6.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.2
wrapt==1.17.2
